requirements-dev.txt
black>=23.3.0,<25.0.0
darglint>=1.8.1,<2.0.0
isort>=5.10.1,<6.0.0
mypy>=1.4.1,<2.0.0
nox==2024.3.2
pycodestyle>=2.8.0,<3.0.0
pydocstyle>=6.1.1,<7.0.0
pyfakefs>=5.0.0,<5.4.0
pylama>=8.4.0,<9.0.0
pylint>=3.0.0,<4.0.0
pytest-asyncio>=0.17.0,<1.0.0
pytest-cov>=3.0.0,<5.0.0
pytest>=7.0.0,<8.0.0
scrapli-replay==2023.7.30
toml>=0.10.2,<1.0.0